cleanup

+++ test/utils/Deployers.sol
@@ -79,7 +79,7 @@ contract Deployers {
 
     // Update this value when you add a new hook flag.
     uint256 hookPermissionCount = 14;
-    uint256 clearAllHookPermisssionsMask = uint256(~uint160(0) >> (hookPermissionCount));
+    uint160 clearAllHookPermisssionsMask = ~uint160(0) >> (hookPermissionCount);
 
     modifier noIsolate() {
         if (msg.sender != address(this)) {
